terminal_prep -> termprep, also improved its API
