+++ src/main.rs
@@ -8,20 +8,19 @@
 
 mod application;
 mod colorscheme;
-mod painters;
-mod langs;
-mod utils;
 mod handlers;
+mod langs;
+mod painters;
+mod utils;
+use smokey;
 
-use std::panic;
 use std::time::Duration;
 use std::{fs::File, io::stdout};
 
 use application::{App, Screen};
 use colorscheme::Theme;
-use crossterm::{execute, style::Print};
-use utils::terminal_prep;
 use handlers::key_handle;
+use utils::termprep;
 
 #[macro_use]
 extern crate log;
@@ -32,21 +31,6 @@ use crossterm::event::{poll, read, Event as CEvent};
 
 use tui::{backend::CrosstermBackend, Terminal};
 
-/// In case of panic restores terminal before program terminates
-fn panic_hook(panic_info: &panic::PanicInfo) {
-    terminal_prep::cleanup_terminal();
-    let msg = match panic_info.payload().downcast_ref::<String>() {
-        Some(s) => format!("p! {}", s),
-        None => match panic_info.payload().downcast_ref::<&str>() {
-            Some(s) => format!("oof! {}", s),
-            None => String::from("weird panic hook"),
-        },
-    };
-    let location = panic_info.location().unwrap();
-    let mut sout = stdout();
-    execute!(sout, Print(format!("{}\n{}\n", msg, location))).unwrap();
-}
-
 fn main() -> crossterm::Result<()> {
     WriteLogger::init(
         LevelFilter::Debug,
@@ -54,8 +38,8 @@ fn main() -> crossterm::Result<()> {
         File::create("smokey.log").unwrap(),
     )
     .expect("logger init went fine");
-    terminal_prep::init_terminal();
-    panic::set_hook(Box::new(|info| panic_hook(info)));
+
+    termprep::init();
 
     #[allow(unused_mut)]
     let mut sout = stdout();
@@ -72,7 +56,6 @@ fn main() -> crossterm::Result<()> {
     app.screen = Screen::Test;
 
     while !app.should_quit {
-
         // draw to terminal using current painter
         (app.painter)(&mut terminal, &mut app);
 
@@ -84,9 +67,9 @@ fn main() -> crossterm::Result<()> {
             if let CEvent::Key(event) = read {
                 key_handle(event, &mut app, &theme);
             }
-        } 
+        }
     }
 
-    terminal_prep::cleanup_terminal();
+    termprep::shutdown();
     Ok(())
 }
